backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
ortools==9.8.3296
groq==0.4.2
python-multipart==0.0.9
reportlab==4.0.9
openpyxl==3.1.2
